+++ client/pages/content.js
@@ -1,47 +1,12 @@
 import React, { useEffect, useState } from 'react';
-import axios from 'axios';
 import Head from 'next/head';
-import Category from '../src/components/Category';
 import Image from 'next/image';
 import pic from '../public/honeycomb.png';
 import Best from '../src/components/Best';
 import Thumbnail from '../src/components/Thumbnail';
+import Category from '../src/components/Category';
 
 export default function Content() {
-    const [postList, setPostList] = useState([]); // []
-    const [initData, setInitData] = useState([]);
-    // const [items, setItems] = useState(10);
-    // const [preItems, setPreItems] = useState(0);
-    const [isClick, setClick] = useState(false);
-
-    const categoryHandler = (e) => {
-        if (e.target.innerText === '전체') {
-            setPostList(initData);
-            return;
-        } else {
-            const filteredData = initData.filter((el) => {
-                return el.category === e.target.innerText;
-            });
-            setPostList(filteredData);
-        }
-    };
-
-    const clickHandler = () => {
-        setClick(!isClick);
-    };
-
-    const getPostsData = () => {
-        axios.get(process.env.NEXT_PUBLIC_URL + '/post').then((res) => {
-            // const result = res.data.post.slice(preItems, items);
-            setPostList(res.data.data);
-            setInitData(res.data.data);
-        });
-    };
-
-    useEffect(() => {
-        getPostsData();
-    }, []);
-
     return (
         <>
             <Head>
@@ -51,9 +16,9 @@ export default function Content() {
             <div>
                 <div className="content">
                     <div className="best_content_container">
-                        <Category isClick={isClick} clickHandler={clickHandler} categoryHandler={categoryHandler} />
+                        <nav className="nav_area"></nav>
                         <Best />
-                        <Thumbnail postList={postList} setPostList={setPostList} />
+                        <Thumbnail />
                     </div>
                 </div>
             </div>
