+++ client/src/components/Category.js
@@ -1,5 +1,25 @@
-export default function Category({ isClick, clickHandler, categoryHandler }) {
+import { useState, useEffect } from 'react';
+
+export default function Category({ postList, setPostList, initData }) {
+    const [isClick, setClick] = useState(false);
     const categories = ['전체', '건강', '운동', '생활', '동물', '쇼핑', '휴지통'];
+
+    const categoryHandler = (e) => {
+        if (e.target.innerText === '전체') {
+            setPostList(initData);
+            return;
+        } else {
+            const filteredData = initData.filter((el) => {
+                return el.category === e.target.innerText;
+            });
+            setPostList(filteredData);
+        }
+    };
+
+    const clickHandler = () => {
+        setClick(!isClick);
+    };
+
     return (
         <>
             {isClick ? (
