+++ client/styles/Content.css
@@ -33,7 +33,6 @@
 .nav_container {
     height: 30%;
     padding-bottom: 3vw;
-    padding-top: 10vw;
     width: 100%;
     box-sizing: inherit;
     display: block;
@@ -277,8 +276,6 @@
     }
     .nav_container {
         height: 30%;
-        padding-bottom: 3vw;
-        padding-top: 10vw;
         width: 100%;
         box-sizing: inherit;
         display: block;
