+++ client/styles/Search.module.css
@@ -1,7 +1,7 @@
 .search_box {
     padding: 0vw;
     position: absolute;
-    left: 60vw;
+    left: 80vw;
     top: 1vw;
     width: 4vw;
     height: 3vw;
