+++ client/styles/Select.module.css
@@ -1,6 +1,6 @@
 .select_container {
     position: relative;
-    left: 68vw;
+    left: 70vw;
     top: 0vw;
 }
 .select {
