+++ client/styles/Tumbnail.module.css
@@ -7,7 +7,6 @@
     position: relative;
 }
 .search_line {
-    left: 10vw;
     display: flex;
     position: relative;
 }
